10197: Extracted phone number generator method to helper.ts file;
