+++ cypress/cypress-integration/support/helpers.ts
@@ -21,3 +21,16 @@ export function b64toBlob(b64Data, contentType, sliceSize) {
   blob.lastModifiedDate = new Date();
   return blob;
 }
+
+export function generateRandomPhoneNumber(): string {
+  function getRandomNumber(min: number, max: number) {
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  }
+
+  const areaCode = getRandomNumber(100, 999);
+  const firstPart = getRandomNumber(100, 999);
+  const secondPart = getRandomNumber(1000, 9999);
+  const phoneNumber = `+1 (${areaCode}) ${firstPart}-${secondPart}`;
+
+  return phoneNumber;
+}
